Rename QualityItem.java to AgingItem
