+++ Java/src/test/java/com/gildedrose/agingitem/AgingItemTest.java
@@ -1,0 +1,15 @@
+package com.gildedrose.agingitem;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class AgingItemTest {
+
+    @Test
+    public void itemQualityIsAccessible() {
+        AgingItem item = new AgingItem("general", 10, 20);
+        assertEquals(20, item.getQuality());
+    }
+
+}
